requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
fuzzywuzzy==0.18.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
Levenshtein==0.27.3
MarkupSafe==3.0.3
nltk==3.9.2
numpy==2.4.0
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
pytz==2025.2
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
textblob==0.19.0
threadpoolctl==3.6.0
tmdbv3api==1.9.0
tqdm==4.67.1
tzdata==2025.3
urllib3==2.6.2
vaderSentiment==3.3.2
Werkzeug==3.1.4
